requirements.txt
streamlit==1.38.0
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
chromadb==0.5.3
pypdf2==3.0.1
sentence-transformers==3.1.1
ollama==0.3.3
python-dotenv==1.0.1
langchain-groq==0.2.0
python-dotenv==1.0.1
